+++ .idea/misc.xml
@@ -5,8 +5,12 @@
       <map>
         <entry key="..\:/Users/estudo/Documents/development/android/letswatch/LetsWatch/app/src/main/res/layout/activity_details.xml" value="0.20520833333333333" />
         <entry key="..\:/Users/estudo/Documents/development/android/letswatch/LetsWatch/app/src/main/res/layout/activity_main.xml" value="0.20520833333333333" />
+        <entry key="..\:/Users/estudo/Documents/development/android/letswatch/LetsWatch/app/src/main/res/layout/error_state_layout.xml" value="0.33" />
         <entry key="..\:/Users/estudo/Documents/development/android/letswatch/LetsWatch/app/src/main/res/layout/gallery_fragment.xml" value="0.20520833333333333" />
         <entry key="..\:/Users/estudo/Documents/development/android/letswatch/LetsWatch/app/src/main/res/layout/gallery_item_layout.xml" value="0.20520833333333333" />
+        <entry key="..\:/Users/estudo/Documents/development/android/letswatch/LetsWatch/app/src/main/res/layout/loading_state_layout.xml" value="0.22" />
+        <entry key="..\:/Users/estudo/Documents/development/android/letswatch/LetsWatch/app/src/main/res/layout/main_content.xml" value="0.18697916666666667" />
+        <entry key="..\:/Users/estudo/Documents/development/android/letswatch/LetsWatch/app/src/main/res/menu/menu_principal.xml" value="0.18697916666666667" />
         <entry key="..\:/Users/jdavi/Documents/development/androidProjects/letsWatch/LetsWatch/app/src/main/res/layout/activity_details.xml" value="0.21467391304347827" />
         <entry key="..\:/Users/jdavi/Documents/development/androidProjects/letsWatch/LetsWatch/app/src/main/res/layout/activity_main.xml" value="0.28940217391304346" />
         <entry key="..\:/Users/jdavi/Documents/development/androidProjects/letsWatch/LetsWatch/app/src/main/res/layout/gallery_fragment.xml" value="0.28940217391304346" />
