+++ app/src/main/AndroidManifest.xml
@@ -8,12 +8,13 @@
     <application
         android:name=".LetsApplication"
         android:allowBackup="true"
+        android:fullBackupOnly="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.LetsWatch">
-        <activity android:name=".ui.moviedetails.DetailsActivity" />
+        <activity android:name=".ui.details.DetailsActivity" />
         <activity android:name=".ui.main.MainActivity"
             android:exported="true"
             >
