+++ app/src/main/java/com/jdavifranco/letswatch/di/ViewModelModule.kt
@@ -1,8 +1,8 @@
 package com.jdavifranco.letswatch.di
 
 import com.jdavifranco.letswatch.ui.main.MainActivityViewModel
-import com.jdavifranco.letswatch.ui.moviedetails.DetailsViewModel
-import com.jdavifranco.letswatch.ui.moviegallery.GalleryViewModel
+import com.jdavifranco.letswatch.ui.details.DetailsViewModel
+import com.jdavifranco.letswatch.ui.gallery.GalleryViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
