+++ app/src/main/java/com/jdavifranco/letswatch/ui/details/DetailsViewModel.kt
@@ -1,10 +1,10 @@
-package com.jdavifranco.letswatch.ui.moviedetails
+package com.jdavifranco.letswatch.ui.details
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.jdavifranco.letswatch.datasource.repository.Repository
+import com.jdavifranco.letswatch.repository.Repository
 import com.jdavifranco.letswatch.domain.model.Details
 import kotlinx.coroutines.launch
