+++ app/src/main/res/layout/gallery_fragment.xml
@@ -1,12 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <data>
         <variable
             name="viewModel"
-            type="com.jdavifranco.letswatch.ui.moviegallery.GalleryViewModel" />
+            type="com.jdavifranco.letswatch.ui.gallery.GalleryViewModel" />
         <variable
             name="movie"
             type="com.jdavifranco.letswatch.domain.model.Movie" />
